Split output_overlays.py into rating-slot and type-cleanup clusters
Splits modules/output_overlays.py (801 lines) into two thematic
sibling modules, keeping the shared naming/reordering/sorting
helpers in the main module.

## What moved

### modules/output_overlays_rating_slots.py (~344 lines, new)

Owns the 12-step rating-slot compaction pipeline:
- Bookkeeping helpers: _is_empty_rating_value, _offset_number,
  _is_ratings_overlay
- Pipeline steps: _clean_template_variables,
  _drop_incomplete_rating_slots, _extract_slot_payloads,
  _distribute_shared_offsets, _reexpand_uniform_vertical_offsets,
  _collapse_uniform_horizontal_offsets,
  _rewrite_legacy_vertical_offsets, _clamp_edge_anchor_offsets,
  _flatten_slot_payloads
- Main entry: prune_rating_template_vars

Plus constants: _EXPLICIT_SLOT_OFFSET_KEYS, _RATINGS_DEFAULT_NAMES.

### modules/output_overlays_type_cleanups.py (~399 lines, new)

Owns the per-overlay-type post-processing cleanup pass:
- Shared helpers: _drop_none_values, _elide_default_builder_level,
  _coerce_string_bool
- Per-type cleaners: _cleanup_resolution_overlay,
  _cleanup_commonsense_overlay, _cleanup_episode_info_overlay,
  _cleanup_languages_overlay, _cleanup_aspect_video_format_overlay
- Main entry: apply_per_overlay_type_cleanup

Plus all constants: DEFAULT_LANGUAGE_FLAG_CODES/WEIGHTS,
_COMMONSENSE_ALIASES (moved here since both overlay_lookup_name
and _cleanup_commonsense_overlay need it), the four *_DEFAULTS sets,
_RESOLUTION_EDITION_* allowlists, and the _OVERLAY_TYPE_CLEANUPS
dispatch table.

## modules/output_overlays.py: 801 -> 217 (-584 / -72.9%)

The 217-line file now contains:
- Rewritten module docstring pointing at the two siblings
- 3 re-export blocks
- Three functions that stay: overlay_lookup_name,
  reorder_rating_template_vars, sort_overlay_entries
- The _RATING_TEMPLATE_VAR_ORDER tuple (used only by
  reorder_rating_template_vars)

## Backward compatibility

output_overlays.py re-exports every moved name.  The only external
importer (modules/output_overlay_builder.py) uses 5 public functions
that all continue to work unchanged.  Zero test-side references
(verified via grep).

## Verification

AST-verified: all 25 moved functions byte-for-byte identical to
origin/develop.  Full test suite passes (0 failures).  Ruff +
Black clean.
